requirements.txt
automium==0.2.6
automium-web==0.1.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
Django>=2.0.2
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Flask==0.12.2
httpie==0.9.9
idna==2.6
itsdangerous==0.24
Jinja2==2.10
legacy==0.1.6
Library==0.0.0
Markdown==2.6.11
MarkupSafe==1.0
Pillow==5.0.0
Pygments==2.2.0
PyJWT==1.5.3
pytz==2018.3
requests==2.18.4
router==0.1
urllib3==1.22
Werkzeug==0.14.1
psycopg2==2.7.4
